balancer: fix SubConn embedding requirement to not recommend a nil panic hazard
